Make string key access fallback to atom key access
With this change `Simple.Json.write` and `Simple.Json.read` works round-trip. Without this change, it always encodes record keys as atoms, and it always reads record keys by string, so all record accesses fail. It feels like a safe default to first check exactly what was asked for, and then fall back to atom access. Also, it's unlikely that we'll do record accesses on user input field names, so atom leaks should not be an issue.